third_party/WebKit/LayoutTests/pointer-lock/pointerlockelement-null-when-pending-expected.txt

Test pointerLockElement is null when a lock is pending.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.webkitPointerLockElement is null
     Lock.
PASS document.webkitPointerLockElement is null
PASS onpointerlockchange received after: Lock.
PASS document.webkitPointerLockElement is targetDiv1
     Unlock.
PASS document.webkitPointerLockElement is targetDiv1
PASS onpointerlockchange received after: Unlock.
PASS document.webkitPointerLockElement is null
PASS successfullyParsed is true

TEST COMPLETE
